Rename snmp plugin to snmp_legacy
