+++ etc/telegraf.conf
@@ -1393,8 +1393,8 @@
 #   servers = ["http://localhost:8098"]
 
 
-# # Reads oids value from one or many snmp agents
-# [[inputs.snmp]]
+# # DEPRECATED! PLEASE USE inputs.snmp INSTEAD.
+# [[inputs.snmp_legacy]]
 #   ## Use 'oids.txt' file to translate oids to names
 #   ## To generate 'oids.txt' you need to run:
 #   ##   snmptranslate -m all -Tz -On | sed -e 's/"//g' > /tmp/oids.txt
